Forgot detecting how backwards-sexp ignores '
backward-sexp is broken (of course) so it goes before the ' not after, which is the equivalent of up-list, THEN backwards-sexp in a world where emacs doesn't suck.

'(a b c)<POINT>
backwards-sexp
<POINT>'(a b c)

(quote (a b c)<POINT>)
backwards-sexp
(quote <POINT>(a b c))

it makes a little sense because it's assuming you're outside the (quote...) implied by ', but that's stupid because nobody would think that intuitively and it's natural to imagine you're inside the (quote...) if you're right of the closing parentheses of '(...).

but this is how emacs imagines it:

(quote (a b c))<POINT>
backwards-sexp
<POINT>(quote (a b c))

so we need to account for suddenly being outside a randomly appearing (quote (...)) list wrapped around the list we want to sort.